tested cases
2.INPUT
Enter the number of days:756
OUTPUT
No. of years:2
No. of weeks:3
No. of days:5
2.INPUT
Enter the number of days:38
OUTPUT
No. of years:0
No. of weeks:5
No. of days:3
3.INPUT
Enter the number of days:0
OUTPUT
No. of years:0
No. of weeks:0
No. of days:0
4.INPUT
Enter the number of days:-365
OUTPUT
No. of years:-1
No. of weeks:0
No. of days:0
5.INPUT
Enter the number of days:-45
OUTPUT
No. of years:0
No. of weeks:-6
No. of days:-3